Split the VisionManager into VisionManager and ShooterManager
- also move math for vision into the VisionManager class and remove all math methods from the Limelight class
- this commit should not cause any behavior changes
