+++ src/main/java/frc/robot/Constants.java
@@ -42,6 +42,7 @@ public final class Constants {
 
     // Vision Manager
     public static final int VISION_MANAGER_PERIOD = 1000 / 22; //22Hz
+    public static final int SHOOTER_MANAGER_PERIOD = 1000 / 22; //22Hz
     public static final double SHOOT_TIME_PER_BALL = 0.2; // For Auto
 
     /**
